added folder to separate typescript and javascript components
