+++ src/components/typescript/contributors.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import ContributorType from "../types/ContributorType";
+import ContributorType from "../../types/ContributorType";
 import axios from "axios";
 
 let Contributor = (props: { contributor: ContributorType; }) => {
