+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import * as React from "react";
-import HeaderComponent from "../components/header.jsx";
-import Main from "../components/main.jsx";
+import HeaderComponent from "../components/javascript/header.jsx";
+import Main from "../components/javascript/main.jsx";
 import { SEO } from "../seo";
 
 const IndexPage = () => {
